improve current state reporting disable verbose output unless DEBUG env is set enable the userdata usage path clean up code clean up artifacts

+++ repainting_3d_assets/paint.py
@@ -1,4 +1,7 @@
+import json
+import os
 import shutil
+from pathlib import Path
 
 import torch
 
@@ -18,8 +21,20 @@ from repainting_3d_assets.view_generation.utils import import_config_key
 from repainting_3d_assets.nerf_reconstruction.train_ngp import sync_config, train_nerf
 
 
-def remove_artifacts(save_dir):
-    shutil.rmtree(f"{save_dir}/transforms.json", ignore_errors=True)
+def dump_configs(save_dir, mesh_config, inpaint_config, nerf_config):
+    def dump_one(name, vals):
+        with open(f"{save_dir}/{name}", "w") as fp:
+            json.dump(vals, fp, indent=4)
+
+    dump_one("config_mesh.log", mesh_config)
+    dump_one("config_inpaint.log", inpaint_config)
+    dump_one("config_nerf.log", nerf_config)
+
+
+def remove_intermediates(save_dir):
+    if "DEBUG" in os.environ:
+        return
+    Path(f"{save_dir}/transforms.json").unlink(missing_ok=True)
     shutil.rmtree(f"{save_dir}/dataset", ignore_errors=True)
     shutil.rmtree(f"{save_dir}/vis", ignore_errors=True)
     shutil.rmtree(f"{save_dir}/sil", ignore_errors=True)
@@ -27,7 +42,9 @@ def remove_artifacts(save_dir):
 
 def paint(mesh_config, inpaint_config, nerf_config):
     device = "cuda" if torch.cuda.is_available() else "cpu"
+    save_dir = mesh_config["save_dir"]
 
+    dump_configs(save_dir, mesh_config, inpaint_config, nerf_config)
     nerf_args = sync_config(nerf_config, mesh_config)
 
     meshes = initialize_meshes(inpaint_config, mesh_config, device)
@@ -110,6 +127,7 @@ def paint(mesh_config, inpaint_config, nerf_config):
 
     nerf_args.record_video = True
     train_nerf(nerf_args)
-    nerf_to_mesh(mesh_config, nerf_args.save_snapshot)
 
-    remove_artifacts(mesh_config["save_dir"])
+    nerf_to_mesh(meshes, save_dir, path_nerf_weights=nerf_args.save_snapshot)
+
+    remove_intermediates(save_dir)
